servicio-1/requirements.txt
Flask
APScheduler 
loging
paho-mqtt